+++ patricare-back/src/modules/address/dto/create-adress.dto.ts
@@ -1,0 +1,8 @@
+import { IsString, IsOptional } from 'class-validator';
+
+export class CreateAdresseDto {
+  @IsString() rue: string;
+
+  @IsOptional()
+  villeId?: number;
+}
