+++ patricare-back/src/modules/address/dto/update-address.dto.ts
@@ -1,0 +1,10 @@
+import { IsString, IsOptional } from 'class-validator';
+
+export class UpdateAdresseDto {
+  @IsOptional()
+  @IsString()
+  rue?: string;
+
+  @IsOptional()
+  villeId?: number;
+}
